--- core/model/src/commonMain/kotlin/io/github/bus/StationName.kt
@@ -1,8 +1,0 @@
-package io.github.bus
-
-value class StationName(val name: String) {
-    companion object {
-        val takinoi = StationName("田喜野井")
-        val tsudanuma = StationName("津田沼_津08")
-    }
-}
