--- core/model/src/commonMain/kotlin/io/github/bus/Timetable.kt
@@ -1,12 +1,0 @@
-package io.github.bus
-
-data class Timetable(
-    val stationName: StationName,
-    val dayType: DayType,
-    val rows: List<TimetableRow>,
-)
-
-data class TimetableRow(
-    val hour: Int,
-    val minutes: List<Int>,
-)
